kæmper for det danske land

+++ public/calenderScript.js
@@ -48,6 +48,7 @@ $(document).ready(function () {
                 let data = {title: title, date: new Date(start.getFullYear(),start.getMonth(),start.getDate(), hourInt, minInt), content: content};
 
 
+
                 fetch(url, {
                     method: "POST",
                     body: JSON.stringify(data),
@@ -87,23 +88,29 @@ $(document).ready(function () {
 
         eventClick: function(info) {
             let id = info._id;
-            console.log(id);
-            let url = 'https://shotokankarate.herokuapp.com/api/calender/' + id;
-            console.log(url);
-            fetch(url, {
-                method: "DELETE",
-            })
-                .then(response => {
-                    if (response.status >= 400)
-                        throw new Error(response.status);
-                    else
+            let tekst = info.title;
 
-                    return response.json();
+            if (confirm("vil du slette: "+ tekst )) {
+                let url = 'https://shotokankarate.herokuapp.com/api/calender/' + id;
+                console.log(url);
+                fetch(url, {
+                    method: "DELETE",
                 })
-                .then(resultat => console.log(`Resultat: %o`, resultat))
-                .catch(fejl => console.log('Fejl: ' + fejl));
-        },
+                    .then(response => {
+                        if (response.status >= 400)
+                            throw new Error(response.status);
+                        else
 
+                            return response.json();
+                    })
+                    .then(resultat => console.log(`Resultat: %o`, resultat))
+                    .catch(fejl => console.log('Fejl: ' + fejl));
+
+            } else {
+
+            }
+
+        },
